requirements.txt
alabaster==0.7.3
alembic==0.7.5.post2
Babel==1.3
blinker==1.3
click==4.0
docutils==0.12
Flask==0.10.1
Flask-Bootstrap==3.3.2.1
Flask-Cors==2.0.1
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-Migrate==1.4.0
flask-mongoengine==0.7.1
Flask-Principal==0.4.0
Flask-PyMongo==0.3.1
Flask-Script==2.0.5
Flask-Security==1.7.4
Flask-SQLAlchemy==2.0
Flask-WTF==0.11
gunicorn==19.3.0
itsdangerous==0.24
Jinja2==2.7.3
Mako==1.0.1
MarkupSafe==0.23
mongoengine==0.9.0
passlib==1.6.2
psycopg2==2.6
Pygments==2.0.2
pytz==2015.2
six==1.9.0
snowballstemmer==1.2.0
Sphinx==1.3.1
SQLAlchemy==1.0.4
Werkzeug==0.10.4
WTForms==2.0.2
